+++ bookingapp/src/main/java/com/dentalcura/bookingapp/model/User.java
@@ -1,7 +1,7 @@
 package com.dentalcura.bookingapp.model;
 
 
-public record User(int id, String name, String surname, String email, String password, Boolean admin) {
+public record User(Long id, String name, String surname, String email, String password, Boolean admin) {
 
     @Override
     public String toString() {
